backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
opencv-python-headless==4.10.0.84
numpy==1.26.4
scipy==1.13.1
Pillow==10.4.0
mediapipe==0.10.14
